Working on adding wrappers for Python objects.
I got stuck up on when pulling a value out of the JS VM if the object was a
wrapped Python object. AFAIK, there's no test for "is pointer a PyObject*" so
I was going with using a PySetType object to track refernces. It appears as
though PyCObject does not get hashed on its void* value. The next step will
be to add a type that inherits from PyCObjectType and does the appropriate
hashing.

Hopefully I can get back to this in some spare time over the next few days.