Working on getting the test suite up to snuff.
I slacked off in keeping the test suite up to date as I was working on the core bits. I'm starting to go through and
convert the old test suite to a new experimental type layout. Hopefully the experiment doesn't blow up in my face. So
far I'm kinda digging it. Still trying to think of ways to cut down on the number of keystrokes per test though.