Fixed failing tests.
NaN and Infinity suprisingly both just worked. It was harder to figure out the correct Python representation for both.

Comparison of JS objects to Python objects in Python now works. I need to repeat for JS land as well as add add checks
for comparing wrapped objects. But for now, back to adding more tests.