Fixed some weird issues with methodobjects.
For some reason passing foo.method around does not increment its reference count
as near as I can tell. This lead to objects being deleted when calling from JS
back to python to execute a method object. I added an extra INCREF in
py2js_object if it's wrapping a PyMethod object. I don't think this is the
correct solution but it works.