Change exception types for resource exhaustion.
Out of memory is now signaled by a MemoryError
Too much CPU time is reported as a SystemError

I'm going to punt on the issue with specifying memory usage limits on contexts
when it really refernces the runtime. I don't see a good way to implement it
without quite a bit of work. It'll never be specifiable per context, only the
api might change to avoid confusion.

[#12 state:resolved]