Fixes global handling of "function foo() {}"
JavaScript acts a bit odd in this case in that the body of 'foo' is passed to
JS_Class.add_property but not JS_Class.set_property. For now I added an add_prop
method that special cases when the passed in value is a function and delegates
to set_prop.

While I was at it, I went ahead and added support for deleting global properties
as well. If your global handler has __delitem__ defined, it will now get called
for the JavaScript statement "delete foo;"

Thanks to Riccardo Pelizzi for the bug report.